+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 Note: Can be used with `sfdx plugins:install sfdx-hardis@beta` and docker image `hardisgroupcom/sfdx-hardis@beta`
 
-## [4.15.0] 2023-11-24
-
+- Allow to disable not monitoring checks using **monitoringDisable** config file property, or **MONITORING_DISABLE** env var
 - Add new feature to identify custom metadata (flows) inactive in project
-- **hardis:lint:metadatastatus** : New command to identify custom metadata (Labels and custom permissions) not used in source code
+  - **hardis:lint:metadatastatus** : New command to identify custom metadata (Labels and custom permissions) not used in source code
 - **Rework generate csv file** : generateReportPath and generateCsvFile
+- Update monitoring and slack documentation
+- Fix slack, teams & Azure notifications
 
 ## [4.14.0] 2023-11-23